+++ docs/.vuepress/config.js
@@ -1,6 +1,7 @@
 module.exports = {
   title: "vue-composable",
-  description: "Vue composition-api composable components",
+  description:
+    "Vue composition-api composable components. i18n, validation, pagination, fetch, etc. +30 variable composable functions",
   plugins: [
     "@vuepress/back-to-top",
     [
@@ -11,7 +12,7 @@ module.exports = {
       }
     ]
   ],
-  base: "/vue-composable/",
+  base: process.env.DOCS_BASE || "",
   head: [
     ["meta", { charset: "utf-8" }],
     ["meta", { name: "msapplication-TileColor", content: "#ffffff" }],
@@ -171,7 +172,9 @@ module.exports = {
             ["composable/web/language", "Language"],
             ["composable/web/broadcastChannel", "BroadcastChannel API"],
             ["composable/web/geolocation", "Geolocation API"],
-            ["composable/web/cssVariables", "CSS variables"]
+            ["composable/web/cssVariables", "CSS variables"],
+            ["composable/web/worker", "WebWorker API"],
+            ["composable/web/workerFunction", "WebWorker Function"]
           ]
         },
         {